// testace-app/shared/types.d.ts
export declare enum DifficultyLevel {
    EASY = "easy",
    MEDIUM = "medium",
    HARD = "hard"
}
export interface TestSession {
    _id: string;
    userId: string;
    mode: TestMode;
    questions: string[];
    answers: UserAnswer[];
    startTime: Date;
    endTime?: Date;
    timeLimit?: number;
    score?: number;
    completed: boolean;
    subject?: string;
    difficulty?: DifficultyLevel;
}
export declare enum TestMode {
    PRACTICE = "practice",
    TIMED = "timed",
    DAILY_CHALLENGE = "daily_challenge"
}